combined formatter and uuid scripts
