--- server/utils/jsonToJsonLines.js
@@ -1,16 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-
-function convertJsonToJsonLines(inputFile, outputFile) {
-  const data = JSON.parse(fs.readFileSync(inputFile, 'utf8'));
-
-  const output = data.map((item) => JSON.stringify(item)).join('\n');
-  fs.writeFileSync(outputFile, output, 'utf8');
-
-  console.log(`Converted ${inputFile} to ${outputFile} in JSON Lines format.`);
-}
-
-const inputFile = path.join(__dirname, './EotsCardsUIDs.json');
-const outputFile = path.join(__dirname, './EotSCardsLines.json');
-
-convertJsonToJsonLines(inputFile, outputFile);
